chore: Minor refactoring of file path
